move route to separate class
